+++ src/lib/client/Client.Constants.ts
@@ -1,0 +1,4 @@
+export enum EArgsEncodings {
+    JSON = 0,
+    RESERVED = 0x10000, // All customized encodings should be greater than this value.
+}