+++ src/lib/client/Client.decl.ts
@@ -23,6 +23,31 @@ export interface IChannelEvents extends Shared.IDefaultEvents {
     ['push_message'](msg: Buffer[]): void;
 }
 
+export interface IApiCallResult<T> {
+
+    /**
+     * The extension binary chunks replied from server.
+     * 
+     * @optional only the server who supports `televoke/2.1` would send this field, if necessary.
+     */
+    binChunks?: Buffer[][];
+
+    /**
+     * The API call result.
+     */
+    result: T;
+}
+
+export interface IApiCallOptions {
+
+    /**
+     * The extension binary chunks to be sent to server.
+     *
+     * @optional only the server who supports `televoke/2.1` could recognize this field, and the old implementation will ignore this field.
+     */
+    binChunks?: Array<Buffer | string | Array<Buffer | string>>;
+}
+
 export interface IClient<TApis extends Shared.IObject>
     extends Shared.IEventListener<IChannelEvents>, Pick<
         Shared.IChannelBase<IChannelEvents>,
@@ -51,6 +76,9 @@ export interface IClient<TApis extends Shared.IObject>
     /**
      * Invoke an API.
      *
+     * > This method does not support extension added in `televoke/2.1`.
+     * > To use the extension feature, please use `callApi` instead.
+     *
      * @param name  The API name.
      * @param args  The arguments.
      */
@@ -58,4 +86,16 @@ export interface IClient<TApis extends Shared.IObject>
         name: Shared.IfIsFn<TApis[TName], TName>,
         ...args: TApis[TName] extends Shared.IFn<infer TArgs, any> ? TArgs : never
     ): Promise<TApis[TName] extends Shared.IFn<any[], infer TRet> ? TRet : never>;
+
+    /**
+     * Invoke an API.
+     *
+     * @param name  The API name.
+     * @param args  The arguments.
+     */
+    callApi<TName extends keyof TApis>(
+        name: Shared.IfIsFn<TApis[TName], TName>,
+        args: TApis[TName] extends Shared.IFn<infer TArgs, any> ? TArgs : never,
+        opts?: IApiCallOptions,
+    ): Promise<TApis[TName] extends Shared.IFn<any[], infer TRet> ? IApiCallResult<TRet> : never>;
 }